Very large commit with alot of changes:
- Moved everything from TRItemStringHandler to TRItemProcesser.

- The cache has now been completely removed. This has been done for
multiple reasons:
* I found that it was not very efficient since permissions always had to
be checked in combination to the cache.
* It was a very complicated part of the plugin that was such a mess that
it was very hard to work with it.

- Renamed TRCacheItem to TRItemProcesser
- Moved list with groups from TRNoItem to TRItemProcesser
- Made processItemString go through groups and check for EE and IC2 item
names.
- The limiter, Item Disabler and Limited Creative now follow the same
rules in the config files.
- (Hopefully) fixed all problems with a single id not blocking all items
- (Hopefully) fixed all problems with a single id with data value 0
blocking other data values as well.
- Added TRException (for use with processItemString
- Made processItemString throw TRExceptions when an error occurred, so
the classes that use it can catch the error and give the user a more
informative error message without the need of giving another parameter
to processItemString.

- Made TRCharge extend TRItem

- Generalized config warnings more, so the format of the errors should
be generally the same across the different classes.

- Removed some unused classes that were there for a possible replacement
of the cache. Since I removed it now, they were no longer necessary.
- Removed old SafeZone class that was still in the project (but excluded
from the build path)