Working on a new improved logfilter: - Filter console only, logs only or both. - Filter on CONTAINS, ENDS_WITH, STARTS_WITH, EQUALS and REGEX. - Filter case sensitive - Filter forge logs - Console enhancer, pretty up/shorten the console messages.
Command Improvements:
- Improved openalc, clearalc and openinv.

I'm releasing this version because of serious bugs in older versions of
TekkitRestrict. It contains a lot of half working features.
I hope someone is able to test this version thoroughly to tell me what
works and doesn't work.

Please ignore the "Delete your config to fix this warning" messages.
That will probably not fix the problem. I recommend deleting all config
files if you want to test this version. If you want to use this version,
do not delete any config files. This will prevent confusion for not
working config options.